Strip the client block down to latency only.
